Protect against bad metadata
fseek() requires integer, so make it such

+++ zp-core/exif/makers/sanyo.php
@@ -152,7 +152,7 @@ function parseSanyo($block, &$result, $seek, $globalOffset) {
 			$value = bin2hex($value);
 			if ($intel == 1)
 				$value = intel2Moto($value);
-			$v = fseek($seek, $globalOffset + hexdec($value));	//offsets are from TIFF header which is 12 bytes from the start of the file
+			$v = fseek($seek, (int) ($globalOffset + hexdec($value))); //offsets are from TIFF header which is 12 bytes from the start of the file
 			if ($tag != 0) {
 				$data = fread($seek, $bytesofdata);
 			} else if ($v == -1) {
